+++ src/NarrativeCharts/NarrativeChartData.cs
@@ -5,8 +5,8 @@ namespace NarrativeCharts;
 public class NarrativeChartData
 {
 	public Dictionary<Character, Hex> Colors { get; set; } = [];
-	public SortedList<int, NarrativeEvent> Events { get; set; } = [];
+	public SortedList<float, NarrativeEvent> Events { get; set; } = [];
 	public string Name { get; set; } = "";
-	public Dictionary<Character, SortedList<int, NarrativePoint>> Points { get; set; } = [];
+	public Dictionary<Character, SortedList<float, NarrativePoint>> Points { get; set; } = [];
 	public Dictionary<Location, int> YIndexes { get; set; } = [];
 }
